pyproject.toml
[project]
name = "agent-study-typer-demo"
version = "0.1.0"
description = "Typer CLI demo for an agent-like terminal chat loop"
requires-python = ">=3.10"
dependencies = [
    "aiohttp",
    "anthropic",
    "google-genai",
    "openai",
    "prompt-toolkit",
    "python-dotenv",
    "rich",
    "typer",
]
